+++ src/popup.js
@@ -3,6 +3,9 @@
 (() => {
   const CONFIG = globalThis.SV_SETTINGS;
 
+  /** @type {HTMLElement} */ (document.getElementById('version')).textContent =
+    `v${chrome.runtime.getManifest().version}`;
+
   /** @type {HTMLElement} */ (document.getElementById('options')).addEventListener('click', (event) => {
     event.preventDefault();
     chrome.runtime.openOptionsPage();